renaming and additional test cases
